+++ packages/ratel_generator/lib/src/injected_types.dart
@@ -3,3 +3,7 @@ import 'package:source_gen/source_gen.dart';
 const requestContextChecker = TypeChecker.fromUrl(
   'package:ratel/core/request_context.dart#RequestContext',
 );
+
+const multipartChecker = TypeChecker.fromUrl(
+  'package:ratel/http/multipart_data.dart#MultipartData',
+);