+++ packages/ratel_generator/test/support/stub_packages.dart
@@ -2,6 +2,7 @@ const stubPackages = <String, String>{
   'ratel|lib/annotations/annotations.dart': _annotations,
   'ratel|lib/core/request_context.dart': _requestContext,
   'ratel|lib/http/handler.dart': _handler,
+  'ratel|lib/http/multipart_data.dart': _multipartData,
   'ratel_orm|lib/annotations.dart': _ormAnnotations,
   'ratel_orm|lib/ratel_orm.dart': _orm,
 };
@@ -125,4 +126,10 @@ class Column {
 
 const _orm = '''
 export 'annotations.dart';
+''';
+
+const _multipartData = '''
+class MultipartData {
+  const MultipartData();
+}
 ''';